--- a/src/Greens_functions.jl
+++ b/src/Greens_functions.jl
@@ -123,7 +123,7 @@ julia> dGdT2s = dG_o_dT2s_x_T2s_superlorentzian((t-τ)/T2s)/T2s
 15253.095033670965
 ```
 """
-dG_o_dT2s_x_T2s_superlorentzian(κ) = quadgk(ζ -> exp(-κ^2 * (3ζ^2 - 1)^2 / 8) * (κ^2 * (3ζ^2 - 1)^2 / 4), 0, sqrt(1/3), 1, order = 100)[1]
+dG_o_dT2s_x_T2s_superlorentzian(κ) = quadgk(ζ -> exp(-κ^2 * (3ζ^2 - 1)^2 / 8) * (κ^2 * (3ζ^2 - 1)^2 / 4), 0, 1/√3, 1, order = 100)[1]
 
 
 """