--- a/src/MatrixExp_Hamiltonians.jl
+++ b/src/MatrixExp_Hamiltonians.jl
@@ -318,7 +318,7 @@ function propagator_linear_inversion_pulse(ω1, T, B1, R2s, _, dR2sdB1, grad_typ
 end
 
 function z_rotation_propagator(rfphase_increment, _)
-	(sϕ,cϕ) = sincos(rfphase_increment)
+    sϕ, cϕ = sincos(rfphase_increment)
     u_rot = @SMatrix [cϕ -sϕ 0 0 0 0;
                       sϕ  cϕ 0 0 0 0;
                        0   0 1 0 0 0;
@@ -329,7 +329,7 @@ function z_rotation_propagator(rfphase_increment, _)
 end
 
 function z_rotation_propagator(rfphase_increment, ::grad_param)
-	(sϕ,cϕ) = sincos(rfphase_increment)
+    sϕ, cϕ = sincos(rfphase_increment)
     u_rot = @SMatrix [cϕ -sϕ 0 0 0  0   0 0 0 0 0;
                       sϕ  cϕ 0 0 0  0   0 0 0 0 0;
                       0    0 1 0 0  0   0 0 0 0 0;
@@ -345,19 +345,19 @@ function z_rotation_propagator(rfphase_increment, ::grad_param)
 end
 
 function xs_destructor(_)
-	Diagonal(SVector{6}(1,1,1,0,1,1))
+    Diagonal(SVector{6}(1,1,1,0,1,1))
 end
 
 function xs_destructor(::grad_param)
-	Diagonal(SVector{11}(1,1,1,0,1,1,1,1,0,1,1))
+    Diagonal(SVector{11}(1,1,1,0,1,1,1,1,0,1,1))
 end
 
-@inline function A0(_)
-	Diagonal(SVector{6}(1,1,1,1,1,0))
+function A0(_)
+    Diagonal(SVector{6}(1,1,1,1,1,0))
 end
 
 function A0(::grad_param)
-	Diagonal(SVector{11}(1,1,1,1,1,1,1,1,1,1,0))
+    Diagonal(SVector{11}(1,1,1,1,1,1,1,1,1,1,0))
 end
 
 function C(_)