* Using `Diagonal{SVector{N}}` instead of `SMatrix{N,N}` where applicable.
* Converted some `exp(1i*x::Real)` expressions to `cis(x)`.