@jecs 2024-09-09
New `test` output:
```
Time to pre-calculate the interpolation:  1.206 ms (273 allocations: 30.06 KiB)
Time to evaluate the original     super-Lorenzian Green's function:  2.083 μs (2 allocations: 32 bytes)
Time to evaluate the interpolated super-Lorenzian Green's function:  237.309 ns (2 allocations: 32 bytes)
Test Summary:        | Pass  Total   Time
Greens_Interpolation |  103    103  15.3s
Test Summary:             | Pass  Total   Time
apply_hamiltonian_gbloch! |   50     50  33.3s
Test Summary:                              | Pass  Total   Time
apply_hamiltonian_gbloch! shaped RF pulses |    9      9  14.9s
Test Summary:                                | Pass  Total   Time
apply_hamiltonian_gbloch! ω₀-sweep RF pulses |    8      8  11.4s
Test Summary:                     | Pass  Total  Time
apply_hamiltonian_freeprecession! |   50     50  0.6s
Test Summary:                             | Pass  Total   Time
apply_hamiltonian_graham_superlorentzian! |   10     10  33.8s
Test Summary:          | Pass  Total     Time
Graham_saturation_rate |    2      2  1m24.8s
Test Summary:           | Pass  Total   Time
apply_hamiltonian_sled! |   10     10  19.3s
Time to pre-compute saturation:  26.763369 seconds (673.64 M allocations: 63.400 GiB, 14.15% gc time, 75.77% compilation time)
Time to solve the full gene. Bloch IDE for 100us π-pulse:  5.486 ms (30775 allocations: 12.07 MiB)
Time to solve the linear approximation for 100us π-pulse:  1.454 μs (6 allocations: 720 bytes)
Test Summary:        | Pass  Total   Time
Linear_Approximation |    9      9  43.9s
Test Summary:                       | Pass  Total   Time
apply_hamiltonian_gbloch! gradients |   35     35  20.0s
Test Summary:                                 | Pass  Total   Time
apply_hamiltonian_gbloch! shaped-RF gradients |   34     34  11.4s
Test Summary:                                   | Pass  Total   Time
apply_hamiltonian_gbloch! ω₀-sweep RF gradients |   36     36  19.3s
Test Summary:                               | Pass  Total  Time
apply_hamiltonian_freeprecession! gradients |   36     36  2.7s
Test Summary:           | Pass  Total     Time
Linear_Approx gradients |   12     12  1m52.0s
Test Summary:                                       | Pass  Total  Time
apply_hamiltonian_graham_superlorentzian! gradients |   34     34  3.6s
Test Summary:  | Pass  Total   Time
inversionpulse |   60     60  23.6s
Test Summary: | Pass  Total   Time
cf. Solvers   |    4      4  38.5s
Test Summary:     | Pass  Total   Time
Solvers_Gradients |    9      9  28.6s
Test Summary:     | Pass  Total   Time
MultipleInversion |    6      6  22.2s
Test Summary: | Pass  Total     Time
OCT Gradients |    7      7  1m31.1s
Test Summary: | Pass  Total   Time
NLLS          |   27     27  52.6s
     Testing MRIgeneralizedBloch tests passed 
 ```

Old `test` output:
```
Time to pre-calculate the interpolation:  1.288 ms (273 allocations: 30.06 KiB)
Time to evaluate the original     super-Lorenzian Green's function:  2.074 μs (2 allocations: 32 bytes)
Time to evaluate the interpolated super-Lorenzian Green's function:  237.827 ns (2 allocations: 32 bytes)
Test Summary:        | Pass  Total   Time
Greens_Interpolation |  103    103  15.6s
Test Summary:             | Pass  Total   Time
apply_hamiltonian_gbloch! |   50     50  35.4s
Test Summary:                              | Pass  Total   Time
apply_hamiltonian_gbloch! shaped RF pulses |    9      9  15.8s
Test Summary:                                | Pass  Total   Time
apply_hamiltonian_gbloch! ω₀-sweep RF pulses |    8      8  12.3s
Test Summary:                     | Pass  Total  Time
apply_hamiltonian_freeprecession! |   50     50  0.7s
Test Summary:                             | Pass  Total   Time
apply_hamiltonian_graham_superlorentzian! |   10     10  35.8s
Test Summary:          | Pass  Total     Time
Graham_saturation_rate |    2      2  1m27.6s
Test Summary:           | Pass  Total   Time
apply_hamiltonian_sled! |   10     10  19.5s
Time to pre-compute saturation:  27.290919 seconds (673.64 M allocations: 63.400 GiB, 13.49% gc time, 71.50% compilation time)
Time to solve the full gene. Bloch IDE for 100us π-pulse:  5.513 ms (30775 allocations: 12.07 MiB)
Time to solve the linear approximation for 100us π-pulse:  1.488 μs (6 allocations: 720 bytes)
Test Summary:        | Pass  Total   Time
Linear_Approximation |    9      9  44.3s
Test Summary:                       | Pass  Total   Time
apply_hamiltonian_gbloch! gradients |   35     35  20.2s
Test Summary:                                 | Pass  Total   Time
apply_hamiltonian_gbloch! shaped-RF gradients |   34     34  11.8s
Test Summary:                                   | Pass  Total   Time
apply_hamiltonian_gbloch! ω₀-sweep RF gradients |   36     36  19.7s
Test Summary:                               | Pass  Total  Time
apply_hamiltonian_freeprecession! gradients |   36     36  2.7s
Test Summary:           | Pass  Total     Time
Linear_Approx gradients |   12     12  1m51.5s
Test Summary:                                       | Pass  Total  Time
apply_hamiltonian_graham_superlorentzian! gradients |   34     34  3.8s
Test Summary:  | Pass  Total   Time
inversionpulse |   60     60  24.3s
Test Summary: | Pass  Total   Time
cf. Solvers   |    4      4  39.3s
Test Summary:     | Pass  Total   Time
Solvers_Gradients |    9      9  30.0s
Test Summary:     | Pass  Total   Time
MultipleInversion |    6      6  22.9s
Test Summary: | Pass  Total     Time
OCT Gradients |    7      7  1m35.1s
Test Summary: | Pass  Total   Time
NLLS          |   27     27  49.5s
     Testing MRIgeneralizedBloch tests passed 
```

So about ~10% faster here and there.