@jecs 2024-09-09
@JakobAsslaender Should be good now.

`grep -Rnw . -e "√" ` returns 0 lines.

`grep -Rnw . -e "\t"` returns 0 lines.